# doc/swagger/rapid_loyalty.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rapid Loyalty API",
    "version": "1.0",
    "contact": {
      "name": "Cong Le",
      "url": "https://github.com/ltcong1411",
      "email": "ltcong1411@gmail.com"
    }
  },
  "tags": [
    {
      "name": "RapidLoyalty"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/customers": {
      "post": {
        "summary": "Create a new customer.",
        "description": "Use this API to create a new customer.",
        "operationId": "RapidLoyalty_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateCustomerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateCustomerRequest"
            }
          }
        ],
        "tags": [
          "RapidLoyalty"
        ]
      }
    },
    "/v1/customers/{barcode}": {
      "get": {
        "summary": "Get a customer by barcode.",
        "description": "Use this API to get a customer by barcode.",
        "operationId": "RapidLoyalty_GetCustomerByBarcode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetCustomerByBarcodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "RapidLoyalty"
        ]
      }
    },
    "/v1/memberships/{barcode}": {
      "get": {
        "summary": "Get a membership by barcode.",
        "description": "Use this API to get a membership by barcode.",
        "operationId": "RapidLoyalty_GetMembershipByBarcode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetMembershipByBarcodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "RapidLoyalty"
        ]
      }
    }
  },
  "definitions": {
    "pbCreateCustomerRequest": {
      "type": "object",
      "properties": {
        "idempotentKey": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      }
    },
    "pbCreateCustomerResponse": {
      "type": "object",
      "properties": {
        "customer": {
          "$ref": "#/definitions/pbCustomer"
        }
      }
    },
    "pbCustomer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "barcode": {
          "type": "string"
        }
      }
    },
    "pbGetCustomerByBarcodeResponse": {
      "type": "object",
      "properties": {
        "customer": {
          "$ref": "#/definitions/pbCustomer"
        }
      }
    },
    "pbGetMembershipByBarcodeResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "customerId": {
          "type": "string"
        },
        "currentTier": {
          "type": "string"
        },
        "availablePoints": {
          "type": "integer",
          "format": "int32"
        },
        "totalPoints": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}